added more events

+++ src/general/State.java
@@ -4,5 +4,7 @@ import java.util.Observable;
 
 public class State extends Observable {
 	boolean isNotRunning = true;
-	int currentTime;
+	double currentTime;
+
+
 }
